+++ src/config.cpp
@@ -391,6 +391,7 @@ bool AdvancedConfig::saveFile() {
   doc[PARAM_HW_PUSH_INTERVAL_INFLUX] = this->getPushIntervalInflux();
   doc[PARAM_HW_PUSH_INTERVAL_MQTT] = this->getPushIntervalMqtt();
   doc[PARAM_HW_TEMPSENSOR_RESOLUTION] = this->getTempSensorResolution();  
+  doc[PARAM_HW_IGNORE_LOW_ANGLES] = this->isIgnoreLowAnges();
 
 #if LOG_LEVEL == 6 && !defined(DISABLE_LOGGING)
   serializeJson(doc, Serial);
@@ -479,6 +480,8 @@ bool AdvancedConfig::loadFile() {
     this->setPushIntervalMqtt(doc[PARAM_HW_PUSH_INTERVAL_MQTT].as<int>());
   if (!doc[PARAM_HW_TEMPSENSOR_RESOLUTION].isNull())
     this->setTempSensorResolution(doc[PARAM_HW_TEMPSENSOR_RESOLUTION].as<int>());
+  if (!doc[PARAM_HW_IGNORE_LOW_ANGLES].isNull())
+    setIgnoreLowAnges(doc[PARAM_HW_IGNORE_LOW_ANGLES].as<bool>());
 
   Log.notice(F("CFG : Configuration file " CFG_HW_FILENAME " loaded." CR));
   return true;