+++ src/resources.hpp
@@ -90,6 +90,7 @@ SOFTWARE.
 #define PARAM_HW_PUSH_INTERVAL_HTTP3 "int-http3"
 #define PARAM_HW_PUSH_INTERVAL_INFLUX "int-influx"
 #define PARAM_HW_PUSH_INTERVAL_MQTT "int-mqtt"
+#define PARAM_HW_IGNORE_LOW_ANGLES "ignore-low-angles"
 #define PARAM_FORMAT_HTTP1 "http-1"
 #define PARAM_FORMAT_HTTP2 "http-2"
 #define PARAM_FORMAT_HTTP3 "http-3"